Move filter into it's own folder
