+++ CefSharp.Example/CefSharp.Example.csproj
@@ -74,7 +74,7 @@
     <Compile Include="DragHandler.cs" />
     <Compile Include="FlashResourceHandler.cs" />
     <Compile Include="FlashResourceHandlerFactory.cs" />
-    <Compile Include="ResponseFilter.cs" />
+    <Compile Include="Filters\ResponseFilter.cs" />
     <Compile Include="ScriptedMethodsBoundObject.cs" />
     <Compile Include="ExceptionTestBoundObject.cs" />
     <Compile Include="JsDialogHandler.cs" />
