+++ CefSharp.Example/RequestHandler.cs
@@ -3,6 +3,7 @@
 // Use of this source code is governed by a BSD-style license that can be found in the LICENSE file.
 
 using System;
+using CefSharp.Example.Filters;
 
 namespace CefSharp.Example
 {
